requirements.txt
uvicorn
fastapi
pydantic
motor
python-dotenv
pymongo
bson